backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "yfinance>=0.2.65",
]
